server/requirements.txt
fastapi==0.115.7
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
PyJWT==2.10.1

